Enable hot reloading react components during development
