+++ frontend/package.json
@@ -27,6 +27,7 @@
         "socket.io-client": "^2.3.0"
     },
     "devDependencies": {
+        "@babel/core": "^7.10.5",
         "@hot-loader/react-dom": "^16.13.0",
         "@ianwalter/exclude-assets-plugin": "^1.0.0",
         "@types/codemirror": "0.0.96",
@@ -42,6 +43,7 @@
         "@typescript-eslint/eslint-plugin-tslint": "^3.6.1",
         "@typescript-eslint/parser": "^3.6.1",
         "awesome-typescript-loader": "^5.2.1",
+        "babel-loader": "^8.1.0",
         "clean-webpack-plugin": "^3.0.0",
         "copy-webpack-plugin": "^5.0.3",
         "css-loader": "^3.6.0",
